requirements.txt
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
torch-geometric==2.6.1
rdkit-pypi
PyTDC
transformers
pandas
numpy
scikit-learn
matplotlib
